+++ src/main/java/com/prowidesoftware/swift/constraints/CountryConstraint.java
@@ -39,7 +39,7 @@ import java.lang.annotation.Target;
 @Target({METHOD, FIELD, PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface CountryConstraint {
-    String message() default "Invalid ISO country code";
+    String message() default "{prowide.invalid.iso.country.code}";
 
     Class<?>[] groups() default {};