+++ src/main/java/com/prowidesoftware/swift/constraints/IbanConstraint.java
@@ -39,7 +39,7 @@ import java.lang.annotation.Target;
 @Target({METHOD, FIELD, PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface IbanConstraint {
-    String message() default "Invalid IBAN number";
+    String message() default "{prowide.invalid.iban.number}";
 
     Class<?>[] groups() default {};