+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Prowide Core - CHANGELOG
 
 #### 10.1.14 - SNAPSHOT
+  * (PW-1812) Updated the narrative resolver, format 2 (used in field 72 for example), to allow empty values as part of the narrative fragment
+  * Updated validators for BIC, country, and currency constraints to utilize keywords for i18n-compatible messages
   * Deprecated unnecessary methods in the SafeXmlUtils class
 
 #### 10.1.13 - December 2023